Rename 'structures' folder to 'models'
