+++ package.json
@@ -15,7 +15,6 @@
         "@sentry/vue": "^6.11.0",
         "@vue/composition-api": "^1.0.0-rc.13",
         "core-js": "^3.8.2",
-        "googleapis": "^39.2.0",
         "marked": "^3.0.0",
         "register-service-worker": "^1.7.2",
         "utds-component-library": "^1.0.9",
