+++ src/models/periods.ts
