+++ src/models/schedule.ts
